+++ NEWS.md
@@ -80,6 +80,8 @@ feols(c(Ozone, Temp) ~ regex("!Day"), airquality)
 
 - `update.(fixest|fixest_multi)` gain the new argument `fml` which can override the `fml.update` argument
 
+- `update.(fixest|fixest_multi)` gain the new argument `use_calling_env` which default to `TRUE`, so that `fixest` objects created within a function can be updated without problem even outside of the function
+
 ## New vignette and data set
 
 - add a new vignette on collinearity